+++ app/Http/Middleware/VerifyCsrfToken.php
@@ -14,6 +14,7 @@ class VerifyCsrfToken extends Middleware
     protected $except = [
         //
         'order',
-        'appload'
+        'appload',
+        'test_config'
     ];
 }