+++ routes/web.php
@@ -23,6 +23,7 @@ Route::get('items/pintoggle/{id}/{ajax?}', 'ItemController@pinToggle')->name('it
 Route::post('order', 'ItemController@setOrder')->name('items.order');
 
 Route::post('appload', 'ItemController@appload')->name('appload');
+Route::post('test_config', 'ItemController@testConfig')->name('test_config');
 
 Route::get('view/{name_view}', function ($name_view) {
     return view('supportedapps.'.$name_view);